+++ packages/isaacscript-common/src/functions/rockAlt.ts
@@ -13,6 +13,7 @@ import {
 import { game } from "../cachedClasses";
 import { DISTANCE_OF_GRID_TILE } from "../constants";
 import { RockAltType } from "../enums/RockAltType";
+import { BACKDROP_TYPE_TO_ROCK_ALT_TYPE } from "../objects/backdropTypeToRockAltType";
 import { spawnEffectWithSeed, spawnNPCWithSeed } from "./entitiesSpecific";
 import { isCollectibleInItemPool } from "./itemPool";
 import {
@@ -22,18 +23,58 @@ import {
   spawnPillWithSeed,
   spawnTrinketWithSeed,
 } from "./pickupsSpecific";
+import { fireProjectilesInCircle } from "./projectiles";
 import { getRandom } from "./random";
 import { getRandomSeed, isRNG, newRNG } from "./rng";
 import { spawnCollectible } from "./spawnCollectible";
 import { repeat } from "./utils";
 import { getRandomVector } from "./vector";
 
+const ROCK_ALT_CHANCES = {
+  Nothing: 0.68,
+  BasicDrop: 0.0967,
+
+  /** Also used for e.g. black hearts from skulls. */
+  Trinket: 0.025,
+
+  Collectible: 0.005,
+} as const;
+
+const POLYP_PROJECTILE_SPEED = 10;
+const POLYP_NUM_PROJECTILES = 6;
+
+/**
+ * Helper function to get the alternate rock type (i.e. urn, mushroom, etc.) that the current room
+ * will have.
+ *
+ * The rock type is based on the backdrop of the room.
+ *
+ * For example, if you change the backdrop of the starting room of the run to `BackdropType.CAVES`,
+ * and then spawn `GridEntityType.ROCK_ALT`, it will be a mushroom instead of an urn. Additionally,
+ * if it is destroyed, it will generate mushroom-appropriate rewards.
+ *
+ * On the other hand, if an urn is spawned first before the backdrop is changed to
+ * `BackdropType.CAVES`, the graphic of the urn will not switch to a mushroom. However, when
+ * destroyed, the urn will still generate mushroom-appropriate rewards.
+ */
+export function getRockAltType(): RockAltType {
+  const room = game.GetRoom();
+  const backdropType = room.GetBackdropType();
+
+  return BACKDROP_TYPE_TO_ROCK_ALT_TYPE[backdropType];
+}
+
 /**
  * Helper function for emulating what happens when a vanilla `GridEntityType.ROCK_ALT` grid entity
  * breaks.
  *
- * Note that most of the time, this function will do nothing, similar to how most of the time, when
- * an individual urn is destroyed, nothing will spawn.
+ * Most of the time, this function will do nothing, similar to how most of the time, when an
+ * individual urn is destroyed, nothing will spawn.
+ *
+ * Note that in vanilla, trinkets will not spawn if they have already been removed from the trinket
+ * pool. This function cannot replicate that behavior because there is no way to check to see if a
+ * trinket is still in the pool. Thus, it will always have a chance to spawn the respective trinket
+ * (e.g. Swallowed Penny from urns).
  *
  * The logic in this function is based on the rewards listed on the wiki:
  * https://bindingofisaacrebirth.fandom.com/wiki/Rocks
@@ -41,6 +82,10 @@ import { getRandomVector } from "./vector";
  * @param position The place to spawn the reward.
  * @param rockAltType The type of reward to spawn. For example, `RockAltType.URN` will have a chance
  *                    at spawning coins and spiders.
+ * @param variant Optional. The variant of the grid entity to emulate. Default is 0, which
+ *                corresponds to a "normal" grid entity or an empty bucket. This only matters when
+ *                spawning the reward for buckets. (Empty buckets have different rewards than full
+ *                buckets.)
  * @param seedOrRNG Optional. The `Seed` or `RNG` object to use. If an `RNG` object is provided, the
  *                  `RNG.Next` method will be called. Default is `getRandomSeed()`. Normally, you
  *                  should pass the `InitSeed` of the grid entity that was broken.
@@ -49,6 +94,7 @@ import { getRandomVector } from "./vector";
 export function spawnRockAltReward(
   position: Vector,
   rockAltType: RockAltType,
+  variant = 0,
   seedOrRNG: Seed | RNG = getRandomSeed(),
 ): boolean {
   const rng = isRNG(seedOrRNG) ? seedOrRNG : newRNG(seedOrRNG);
@@ -71,7 +117,7 @@ export function spawnRockAltReward(
     }
 
     case RockAltType.BUCKET: {
-      return spawnRockAltRewardBucket(position, rng);
+      return spawnRockAltRewardBucket(position, rng, variant);
     }
   }
 }
@@ -80,14 +126,12 @@ function spawnRockAltRewardUrn(position: Vector, rng: RNG): boolean {
   const chance = getRandom(rng);
   let totalChance = 0;
 
-  // Nothing (68.00%).
-  totalChance += 0.68;
+  totalChance += ROCK_ALT_CHANCES.Nothing;
   if (chance < totalChance) {
     return false;
   }
 
-  // One or two random coins (9.44%).
-  totalChance += 0.0944;
+  totalChance += ROCK_ALT_CHANCES.BasicDrop;
   if (chance < totalChance) {
     const numCoinsChance = getRandom(rng);
     const numCoins = numCoinsChance < 0.5 ? 1 : 2;
@@ -100,15 +144,13 @@ function spawnRockAltRewardUrn(position: Vector, rng: RNG): boolean {
     return true;
   }
 
-  // Swallowed Penny (2.50%).
-  totalChance += 0.025;
+  totalChance += ROCK_ALT_CHANCES.Trinket;
   if (chance < totalChance) {
     spawnTrinketWithSeed(TrinketType.SWALLOWED_PENNY, position, rng);
     return true;
   }
 
-  // A Quarter (0.50%).
-  totalChance += 0.005;
+  totalChance += ROCK_ALT_CHANCES.Collectible;
   if (chance < totalChance) {
     const stillInPools = isCollectibleInItemPool(
       CollectibleType.QUARTER,
@@ -122,8 +164,7 @@ function spawnRockAltRewardUrn(position: Vector, rng: RNG): boolean {
     return false;
   }
 
-  // One or two spiders (19.48%). (Since this is the final option, so we don't need to check the
-  // chance.)
+  // Since the detrimental effect is the final option, we don't need to check the chance.
   const numSpidersChance = getRandom(rng);
   const numSpiders = numSpidersChance < 0.5 ? 1 : 2;
   const length = DISTANCE_OF_GRID_TILE * 3;
@@ -143,28 +184,24 @@ function spawnRockAltRewardMushroom(position: Vector, rng: RNG): boolean {
   const chance = getRandom(rng);
   let totalChance = 0;
 
-  // Nothing (68.00%).
-  totalChance += 0.68;
+  totalChance += ROCK_ALT_CHANCES.Nothing;
   if (chance < totalChance) {
     return false;
   }
 
-  // One pill (9.82%).
-  totalChance += 0.0982;
+  totalChance += ROCK_ALT_CHANCES.BasicDrop;
   if (chance < totalChance) {
     spawnPillWithSeed(PillColor.NULL, position, rng);
     return true;
   }
 
-  // Liberty Cap (2.50%).
-  totalChance += 0.025;
+  totalChance += ROCK_ALT_CHANCES.Trinket;
   if (chance < totalChance) {
     spawnTrinketWithSeed(TrinketType.LIBERTY_CAP, position, rng);
     return true;
   }
 
-  // Collectible (0.50%).
-  totalChance += 0.005;
+  totalChance += ROCK_ALT_CHANCES.Collectible;
   if (chance < totalChance) {
     if (roomType === RoomType.SECRET) {
       const wavyCapChance = getRandom(rng);
@@ -211,8 +248,7 @@ function spawnRockAltRewardMushroom(position: Vector, rng: RNG): boolean {
     return false;
   }
 
-  // One poisonous gas effect (19.24%). (Since this is the final option, so we don't need to check
-  // the chance.)
+  // Since the detrimental effect is the final option, we don't need to check the chance.
   spawnEffectWithSeed(EffectVariant.FART, 0, position, rng);
   return true;
 }
@@ -221,28 +257,24 @@ function spawnRockAltRewardSkull(position: Vector, rng: RNG): boolean {
   const chance = getRandom(rng);
   let totalChance = 0;
 
-  // Nothing (68.00%).
-  totalChance += 0.68;
+  totalChance += ROCK_ALT_CHANCES.Nothing;
   if (chance < totalChance) {
     return false;
   }
 
-  // One card / rune (9.52%).
-  totalChance += 0.095;
+  totalChance += ROCK_ALT_CHANCES.BasicDrop;
   if (chance < totalChance) {
     spawnCardWithSeed(Card.NULL, position, rng);
     return true;
   }
 
-  // Black heart (2.50%).
-  totalChance += 0.025;
+  totalChance += ROCK_ALT_CHANCES.Trinket;
   if (chance < totalChance) {
     spawnHeartWithSeed(HeartSubType.BLACK, position, rng);
     return true;
   }
 
-  // Collectible (0.50%).
-  totalChance += 0.005;
+  totalChance += ROCK_ALT_CHANCES.Collectible;
   if (chance < totalChance) {
     const ghostBabyStillInPools = isCollectibleInItemPool(
       CollectibleType.GHOST_BABY,
@@ -275,31 +307,86 @@ function spawnRockAltRewardSkull(position: Vector, rng: RNG): boolean {
     return false;
   }
 
-  // One Host (19.30%). (Since this is the final option, so we don't need to check the chance.)
+  // Since the detrimental effect is the final option, we don't need to check the chance.
   spawnNPCWithSeed(EntityType.HOST, 0, 0, position, rng);
   return true;
 }
 
-function spawnRockAltRewardPolyp(_position: Vector, rng: RNG): boolean {
+function spawnRockAltRewardPolyp(position: Vector, rng: RNG): boolean {
   const chance = getRandom(rng);
   let totalChance = 0;
 
-  // Nothing (68.00%).
-  totalChance += 0.68;
+  totalChance += ROCK_ALT_CHANCES.Nothing;
   if (chance < totalChance) {
     return false;
   }
 
-  // TODO
-  return false;
+  totalChance += ROCK_ALT_CHANCES.BasicDrop;
+  if (chance < totalChance) {
+    spawnHeartWithSeed(HeartSubType.NULL, position, rng);
+    return true;
+  }
+
+  totalChance += ROCK_ALT_CHANCES.Trinket;
+  if (chance < totalChance) {
+    spawnTrinketWithSeed(TrinketType.UMBILICAL_CORD, position, rng);
+    return true;
+  }
+
+  totalChance += ROCK_ALT_CHANCES.Collectible;
+  if (chance < totalChance) {
+    const placentaStillInPools = isCollectibleInItemPool(
+      CollectibleType.PLACENTA,
+      ItemPoolType.BOSS,
+    );
+    const bloodClotStillInPools = isCollectibleInItemPool(
+      CollectibleType.BLOOD_CLOT,
+      ItemPoolType.BOSS,
+    );
+    if (bloodClotStillInPools && placentaStillInPools) {
+      const collectibleChance = getRandom(rng);
+      const collectibleType =
+        collectibleChance < 0.5
+          ? CollectibleType.PLACENTA // 218
+          : CollectibleType.BLOOD_CLOT; // 254
+      spawnCollectible(collectibleType, position, rng);
+      return true;
+    }
+
+    if (bloodClotStillInPools) {
+      spawnCollectible(CollectibleType.MINI_MUSH, position, rng);
+      return true;
+    }
+
+    if (placentaStillInPools) {
+      spawnCollectible(CollectibleType.MAGIC_MUSHROOM, position, rng);
+      return true;
+    }
+
+    return false;
+  }
+
+  // Since the detrimental effect is the final option, we don't need to check the chance.
+  spawnEffectWithSeed(EffectVariant.CREEP_RED, 0, position, rng);
+  fireProjectilesInCircle(
+    undefined,
+    position,
+    POLYP_PROJECTILE_SPEED,
+    POLYP_NUM_PROJECTILES,
+  );
+
+  return true;
 }
 
-function spawnRockAltRewardBucket(_position: Vector, rng: RNG): boolean {
+function spawnRockAltRewardBucket(
+  _position: Vector,
+  rng: RNG,
+  _variant: int,
+): boolean {
   const chance = getRandom(rng);
   let totalChance = 0;
 
-  // Nothing (68.00%).
-  totalChance += 0.68;
+  totalChance += ROCK_ALT_CHANCES.Nothing;
   if (chance < totalChance) {
     return false;
   }